# Backyard-app/openapi.yaml
openapi: 3.0.1
info:
  title: Backyard
  description: This is a Backyard server.
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: pietrusjakub@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
...
servers:
  - url: 'https://backyard.swagger.io/v1'
  - url: 'http://backyard.swagger.io/v1'
...
paths:
  /game:
    get:
      tags:
        - game
      summary: Get games list
      description: Location type values can be provided with comma separated strings
      operationId: getGamesList
      parameters:
        - name: area
          in: query
          description: Location type values that need to be considered for filter
          required: true
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
              enum:
                - court
                - sporthall
        - name: limit
          in: query
          description: The maximum number of results to retrieve
          schema:
            type: integer
        - name: offset
          in: query
          description: The zero-ary offset into the results
          schema:
            type: integer
        - name: sort
          in: query
          description: How to sort the results
          schema:
            type: string
            default: id
            enum:
              - id
              - date
              - price
              - location
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
        '400':
          description: Invalid locationType value
          content: {}
    put:
      tags:
        - game
      summary: Update an existing game
      operationId: updateGame
      requestBody:
        description: Game object that needs to be updated
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Game'
          application/xml:
            schema:
              $ref: '#/components/schemas/Game'
        required: true
      responses:
        '400':
          description: Invalid ID supplied
          content: {}
        '404':
          description: Game not found
          content: {}
        '405':
          description: Validation exception
          content: {}
      x-codegen-request-body-name: body
    post:
      tags:
        - game
      summary: Add a new game
      operationId: addGame
      requestBody:
        description: Game object that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Game'
          application/xml:
            schema:
              $ref: '#/components/schemas/Game'
        required: true
      responses:
        '405':
          description: Invalid input
          content: {}
      x-codegen-request-body-name: body
  '/game/{gameId}':
    get:
      tags:
        - game
      summary: Finds games by gameId
      description: Multiple status values can be provided with comma separated strings
      operationId: getGameById
      parameters:
        - name: gameId
          in: path
          description: Id of game to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
            application/xml:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          description: Invalid ID supplied
          content: {}
        '404':
          description: Game not found
          content: {}
    delete:
      tags:
        - game
      summary: Deletes a game
      operationId: deleteGame
      parameters:
        - name: api_key
          in: header
          schema:
            type: string
        - name: gameId
          in: path
          description: Game id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: Invalid ID supplied
          content: {}
        '404':
          description: Game not found
          content: {}
  '/game/{gameId}/rate':
    post:
      tags:
        - game
      summary: Rate game
      operationId: rateGame
      parameters:
        - name: gameId
          in: path
          description: Id of game to rate
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Rat object that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameRating'
          application/xml:
            schema:
              $ref: '#/components/schemas/GameRating'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameRating'
            application/xml:
              schema:
                $ref: '#/components/schemas/GameRating'
        '405':
          description: Invalid input
          content: {}
      x-codegen-request-body-name: body
  '/game/{gameId}/remove/{userId}':
    delete:
      tags:
        - game
      summary: Remove user from game
      operationId: removeUserFromGame
      parameters:
        - name: gameId
          in: path
          description: ID of game
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: path
          description: ID of user that want to remove from game
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
            application/xml:
              schema:
                $ref: '#/components/schemas/Game'
        '405':
          description: Invalid input
          content: {}
  '/game/{gameId}/uploadImage':
    post:
      tags:
        - game
      summary: Uploads an image
      operationId: uploadFile
      parameters:
        - name: gameId
          in: path
          description: ID of game to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              properties:
                additionalMetadata:
                  type: string
                  description: Additional data to pass to server
                file:
                  type: string
                  description: file to upload
                  format: binary
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  '/game/{gameId}/message':
    get:
      tags:
        - game
      summary: Get messages list
      description: Location type values can be provided with comma separated strings
      operationId: getMessagesList
      parameters:
        - name: gameId
          in: path
          description: ID of game from which messages we want
          required: true
          schema:
            type: integer
            format: int64
        - name: limit
          in: query
          description: The maximum number of results to retrieve
          schema:
            type: integer
        - name: offset
          in: query
          description: The zero-ary offset into the results
          schema:
            type: integer
        - name: sort
          in: query
          description: How to sort the results
          schema:
            type: string
            default: id
            enum:
              - id
              - date
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        '400':
          description: Invalid gameId value
          content: {}
  '/game/{gameId}/message/{userId}':
    post:
      tags:
        - game
      summary: Add a new message
      operationId: addMessage
      parameters:
        - name: gameId
          in: path
          description: Id of game
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: path
          description: Id of user that send message
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Message object that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Message'
          application/xml:
            schema:
              $ref: '#/components/schemas/Message'
        required: true
      responses:
        '405':
          description: Invalid input
          content: {}
      x-codegen-request-body-name: body
  '/game/{gameId}/join/{userId}/{haveABall}':
    post:
      tags:
        - game
      summary: Join player to game
      operationId: joinGame
      parameters:
        - name: gameId
          in: path
          description: ID of game to join
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: path
          description: ID of user that want to join
          required: true
          schema:
            type: integer
            format: int64
        - name: haveABall
          in: path
          description: Boolean if user have a ball
          required: true
          schema:
            type: string
            format: boolean
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
            application/xml:
              schema:
                $ref: '#/components/schemas/Game'
        '405':
          description: Invalid input
          content: {}
...
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
...
    Message:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        content:
          type: string
        date:
          type: string
          format: datetime
      xml:
        name: Message
    Game:
      required:
        - date
        - playersMaxAmount
      type: object
      properties:
        id:
          type: integer
          format: int64
        courtId:
          type: integer
          format: int64
        sportHallId:
          type: integer
          format: int64
        price:
          type: number
          format: float
        date:
          type: string
          format: date-time
        gotBall:
          type: string
          format: boolean
        playersMaxAmount:
          type: integer
          format: int32
        photoUrls:
          type: array
          items:
            type: string
      xml:
        name: Gam
...
    GameRating:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        value:
          type: integer
          format: int32
      xml:
        name: CourtRating